refactor package and remove bad ascii char
